consolidate changes into PragueGasCalculator
